The size of the OuterHeaderRemoval IE was one octet in 3GPP TS 29.244 up to V15.3.0, but it has been changed to two octets since V15.4.0.
The following commit fixes it so that it does not cause an error
even in OuterHeaderRemoval IE with size 1 according to V15.3.0 or
earlier, but when GTPUExternsionHeaderDeletion is called in
OuterHeaderRemoval IE with size 1, v[1] access and index out of
range error will occur.
https://github.com/wmnsk/go-pfcp/commit/78b3fd79723d4ef6e844acb5e31780867c3247f3

This patch fixes OuterHeaderRemoval IE with size 1 to return 0 when
GTPUExternsionHeaderDeletion is called. But, perhaps it would have
been better to make it return an error.